initial transfer to allow c8y server to work.
