+++ gulpfile.js
@@ -30,59 +30,68 @@ const compile = series(
             .pipe(bump())
             .pipe(dest('./'));
     },
-    function buildAngularLibrary() { return ngPackagr.build({ project: './ng-package.json' }) },
-    function separateWebpackBuildSrc() { return fs.copy('./dist/widget-library/fesm5', './dist/bundle-src') },
+    function buildAngularLibrary() {
+        return ngPackagr.ngPackagr()
+            .forProject('./ng-package.json')
+            .withTsConfig('./tsconfig.rt.json')
+            .build()
+            .catch(error => {
+                console.error(error);
+                process.exit(1);
+            });
+    },
+    function separateWebpackBuildSrc() { return fs.copy('./dist/widget-library/fesm5', './dist/bundle-src'); },
     function replaceStylePath() {
         return src('./dist/widget-library/**/*')
             .pipe(replace(/~styles/g, function () {
-                return path.relative(this.file.dirname, './dist/widget-library/styles').replace(/\\/g, '/')
+                return path.relative(this.file.dirname, './dist/widget-library/styles').replace(/\\/g, '/');
             }))
-            .pipe(dest('./dist/widget-library/'))
+            .pipe(dest('./dist/widget-library/'));
     },
-    async function packLibrary() { return execSync("npm pack ./widget-library", { cwd: './dist', stdio: 'inherit' }) }
-)
+    async function packLibrary() { return execSync("npm pack ./widget-library", { cwd: './dist', stdio: 'inherit' }); }
+);
 
 const compileNoBump = series(
-    function buildAngularLibrary() { return ngPackagr.build({ project: './ng-package.json' }) },
-    function separateWebpackBuildSrc() { return fs.copy('./dist/widget-library/fesm5', './dist/bundle-src') },
+    function buildAngularLibrary() { return ngPackagr.build({ project: './ng-package.json' }); },
+    function separateWebpackBuildSrc() { return fs.copy('./dist/widget-library/fesm5', './dist/bundle-src'); },
     function replaceStylePath() {
         return src('./dist/widget-library/**/*')
             .pipe(replace(/~styles/g, function () {
-                return path.relative(this.file.dirname, './dist/widget-library/styles').replace(/\\/g, '/')
+                return path.relative(this.file.dirname, './dist/widget-library/styles').replace(/\\/g, '/');
             }))
-            .pipe(dest('./dist/widget-library/'))
+            .pipe(dest('./dist/widget-library/'));
     },
-    async function packLibrary() { return execSync("npm pack ./widget-library", { cwd: './dist', stdio: 'inherit' }) }
-)
+    async function packLibrary() { return execSync("npm pack ./widget-library", { cwd: './dist', stdio: 'inherit' }); }
+);
 
 /**
  * Create the webpack build include the relvent items and zip it.
  */
 const bundle = series(
-    async function webpackBuild() { return execSync("npx webpack", { stdio: 'inherit' }) },
-    function copyCumulocityJson() { return fs.copy('./cumulocity.json', './dist/widget/cumulocity.json') },
+    async function webpackBuild() { return execSync("npx webpack", { stdio: 'inherit' }); },
+    function copyCumulocityJson() { return fs.copy('./widget-cumulocity.json', './dist/widget/cumulocity.json'); },
     function createZip() {
         const pkgJson = require('./dist/widget-library/package.json');//need bumped version.
         return src('./dist/widget/**/*')
             // Filter out the webpackRuntime chunk, we only need the widget code chunks
             .pipe(filter(file => !/^[a-f0-9]{20}\.js(\.map)?$/.test(file.relative)))
             .pipe(zip(`${pkgJson.name}-${pkgJson.version}.zip`))
-            .pipe(dest('dist/'))
+            .pipe(dest('dist/'));
     }
-)
+);
 
 const bundleRelease = series(
-    async function webpackBuild() { return execSync("npx webpack", { stdio: 'inherit' }) },
-    function copyCumulocityJson() { return fs.copy('./cumulocity.json', './dist/widget/cumulocity.json') },
+    async function webpackBuild() { return execSync("npx webpack", { stdio: 'inherit' }); },
+    function copyCumulocityJson() { return fs.copy('./widget-cumulocity.json', './dist/widget/cumulocity.json'); },
     function createZip() {
         const pkgJson = require('./dist/widget-library/package.json');//need bumped version.
         return src('./dist/widget/**/*')
             // Filter out the webpackRuntime chunk, we only need the widget code chunks
             .pipe(filter(file => !/^[a-f0-9]{20}\.js(\.map)?$/.test(file.relative)))
             .pipe(zip(`${pkgJson.name}-${pkgJson.version}.zip`))
-            .pipe(dest('release/'))
+            .pipe(dest('release/'));
     }
-)
+);
 
 exports.clean = clean;
 exports.build = compile;
