+++ ng-package.json
@@ -1,16 +1,18 @@
 {
-  "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
-  "assets": [
-    "styles/**/*",
-    "assets/**/*"
-],
-  "lib": {
-    "entryFile": "./src/public_api.ts",
-    "umdModuleIds": {
-      "@c8y/ngx-components": "@c8y/ngx-components"
+    "$schema": "./node_modules/ng-packagr/ng-package.schema.json",
+    "assets": [
+        "./styles/**/*",
+        "./src/*/assets/**/*"
+    ],
+    "lib": {
+        "entryFile": "./src/public_api.ts",
+        "umdModuleIds": {
+            "@c8y/ngx-components": "@c8y/ngx-components"
+        },
+        "flatModuleFile": "custom-widget"
     },
-    "flatModuleFile": "custom-widget"
-  },
-  "whitelistedNonPeerDependencies": ["."],
-  "dest": "dist/widget-library"
+    "whitelistedNonPeerDependencies": [
+        "."
+    ],
+    "dest": "dist/widget-library"
 }
