+++ ng1.ts
@@ -1,0 +1,42 @@
+import '@c8y/ng1-modules/core';
+import '@c8y/ng1-modules/imageWidget/cumulocity.json';
+import '@c8y/ng1-modules/assetPropertyWidget/cumulocity.json';
+import '@c8y/ng1-modules/devicePropertyWidget/cumulocity.json';
+import '@c8y/ng1-modules/alarms/cumulocity.json';
+import '@c8y/ng1-modules/assetTable/cumulocity.json';
+import '@c8y/ng1-modules/eventsBinary/cumulocity.json';
+import '@c8y/ng1-modules/devicemanagement-alarmList/cumulocity.json';
+import '@c8y/ng1-modules/deviceSelector/cumulocity.json';
+import '@c8y/ng1-modules/kpi/cumulocity.json';
+import '@c8y/ng1-modules/kpiAdmin/cumulocity.json';
+import '@c8y/ng1-modules/devicemanagement-location/cumulocity.json';
+import '@c8y/ng1-modules/dashboard2/cumulocity.json';
+import '@c8y/ng1-modules/dashboardUI/cumulocity.json';
+import '@c8y/ng1-modules/groupsHierarchy/cumulocity.json';
+import '@c8y/ng1-modules/measurements/cumulocity.json';
+import '@c8y/ng1-modules/map/cumulocity.json';
+import '@c8y/ng1-modules/alarmAssets/cumulocity.json';
+import '@c8y/ng1-modules/deviceControlMessage/cumulocity.json';
+import '@c8y/ng1-modules/deviceControlRelay/cumulocity.json';
+import '@c8y/ng1-modules/deviceControlRelayArray/cumulocity.json';
+import '@c8y/ng1-modules/cockpit-cockpitConfig/cumulocity.json';
+import '@c8y/ng1-modules/cockpit-assetCount/cumulocity.json';
+import '@c8y/ng1-modules/cockpit-alarmRecent/cumulocity.json';
+import '@c8y/ng1-modules/cockpit-reports/cumulocity.json';
+import '@c8y/ng1-modules/cockpit-smartRulesUI/cumulocity.json';
+import '@c8y/ng1-modules/cockpit-dataPointExplorerUI/cumulocity.json';
+import '@c8y/ng1-modules/alarmsEventsExplorer/cumulocity.json';
+import '@c8y/ng1-modules/deviceDatabase4/cumulocity.json';
+import '@c8y/ng1-modules/modbusWidget4/cumulocity.json';
+import '@c8y/ng1-modules/scada/cumulocity.json';
+import '@c8y/ng1-modules/imageWidget/cumulocity.json';
+import '@c8y/ng1-modules/htmlWidget/cumulocity.json';
+import '@c8y/ng1-modules/applicationLinks/cumulocity.json';
+import '@c8y/ng1-modules/quickLinks/cumulocity.json';
+import '@c8y/ng1-modules/helpAndSupport/cumulocity.json';
+import '@c8y/ng1-modules/eventList/cumulocity.json';
+import '@c8y/ng1-modules/export/cumulocity.json';
+import '@c8y/ng1-modules/dataPointTable/cumulocity.json';
+import '@c8y/ng1-modules/switchDisplay/cumulocity.json';
+import '@c8y/ng1-modules/trafficLightWidget/cumulocity.json';
+import '@c8y/ng1-modules/infoGauge/cumulocity.json';
