+++ package.json
@@ -1,19 +1,37 @@
 {
-    "name": "cumulocity-datapoints-chart-widget",
-    "interleave": {
-        "dist\\bundle-src\\custom-widget.js": "cumulocity-datapoints-charting-CustomWidget",
-        "dist/bundle-src/custom-widget.js": "cumulocity-datapoints-charting-CustomWidget"
-    },
-    "version": "1.1.3",
-    "description": "Widget for displaying realtime measurement data in graphs (written by Software AG Global Pre-sales)",
+    "name": "cumulocity-datapoints-charting-widget",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
     "scripts": {
-        "build": "gulp",
-        "release": "gulp release"
+        "start": "c8ycli server",
+        "build": "c8ycli build",
+        "deploy": "c8ycli deploy",
+        "buildRuntime": "gulp",
+        "dev-install": "npm install --force --development"
     },
     "keywords": [],
-    "author": "John Heath - Software AG, Global Pre-sales",
-    "license": "Apache 2.0",
+    "author": "",
+    "license": "ISC",
     "dependencies": {
+        "@angular/animations": "8.2.13",
+        "@angular/cdk": "8.2.3",
+        "@angular/common": "8.2.13",
+        "@angular/compiler": "8.2.13",
+        "@angular/core": "8.2.13",
+        "@angular/forms": "8.2.13",
+        "@angular/platform-browser": "8.2.13",
+        "@angular/platform-browser-dynamic": "8.2.13",
+        "@angular/router": "8.2.13",
+        "@angular/upgrade": "8.2.13",
+        "@c8y/ngx-components": "1009.0.21",
+        "core-js": "^3.4.0",
+        "rxjs": "~6.4.0",
+        "url-search-params-polyfill": "6.0.0",
+        "zone.js": "~0.9.1",
+        "@c8y/ng1-modules": "1009.0.21",
+        "angular": "1.6.9",
+        "@c8y/style": "1009.0.21",
         "@ng-select/ng-select": "^3.7.3",
         "bollinger-bands": "^3.0.1",
         "chart.js": "^2.9.4",
@@ -23,53 +41,45 @@
         "lodash": "^4.17.20",
         "moment": "^2.29.1",
         "moving-averages": "^4.0.5",
-        "ng2-charts": "^2.4.2",
-        "ngx-color-picker": "^11.0.0",
-        "uuid": "^8.3.2"
-    },
-    "peerDependencies": {
-        "@angular/common": "8.2.13",
-        "@angular/core": "8.2.13",
-        "@angular/forms": "8.2.13",
-        "@angular/http": "7.2.15",
-        "@angular/router": "8.2.13",
-        "@c8y/ngx-components": "1006.3.0"
+        "ng2-charts": "^2.4.2"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "0.803.17",
-        "@angular/animations": "8.2.13",
-        "@angular/cdk": "8.2.3",
-        "@angular/common": "8.2.13",
-        "@angular/compiler": "8.2.13",
+        "@angular-devkit/build-angular": "0.803.23",
         "@angular/compiler-cli": "8.2.13",
-        "@angular/core": "8.2.13",
-        "@angular/forms": "8.2.13",
-        "@angular/http": "7.2.15",
-        "@angular/platform-browser": "8.2.13",
-        "@angular/router": "8.2.13",
-        "@c8y/ngx-components": "1006.3.0",
-        "ajv": "^6.9.1",
-        "core-js": "^3.4.0",
-        "css-loader": "^3.5.3",
-        "del": "^5.1.0",
-        "delay": "^4.3.0",
-        "fs-extra": "^9.0.0",
-        "gulp": "^4.0.2",
-        "gulp-bump": "^3.2.0",
-        "gulp-filter": "^6.0.0",
-        "gulp-replace": "^1.0.0",
-        "gulp-zip": "^5.0.1",
-        "ng-packagr": "^9.1.1",
-        "ngx-bootstrap": "5.5.0",
-        "rxjs": "~6.4.0",
-        "rxjs-compat": "^6.5.3",
-        "source-map-loader": "^0.2.4",
+        "@angular/language-service": "8.2.13",
+        "@c8y/cli": "1009.0.21",
+        "caniuse-lite": "1.0.30001237",
+        "css-loader": "3.5.3",
+        "gulp": "4.0.2",
+        "gulp-bump": "3.2.0",
+        "gulp-filter": "6.0.0",
+        "gulp-replace": "1.0.0",
+        "gulp-zip": "5.0.2",
+        "ng-packagr": "^9.1.5",
         "typescript": "3.5.3",
-        "url-loader": "^4.1.0",
-        "webpack": "^4.43.0",
-        "webpack-cli": "^3.3.11",
-        "webpack-dev-middleware": "^4.1.0",
-        "webpack-external-import": "^2.2.3",
-        "zone.js": "~0.9.1"
+        "url-loader": "4.1.1",
+        "webpack": "4.43.0",
+        "webpack-cli": "3.3.11",
+        "webpack-external-import": "2.2.3"
+    },
+    "c8y": {
+        "application": {
+            "name": "cockpit",
+            "contextPath": "cockpit",
+            "key": "cockpit-application-key",
+            "brandingEntry": "styles/index.less",
+            "tabsHorizontal": true,
+            "upgrade": true,
+            "rightDrawer": true,
+            "breadcrumbs": false,
+            "sensorAppOneLink": "http://onelink.to/pca6qe",
+            "sensorPhone": true,
+            "contentSecurityPolicy": "base-uri 'none'; default-src 'self' 'unsafe-inline' http: https: ws: wss:; connect-src 'self' *.billwerk.com http: https: ws: wss:;  script-src 'self' open.mapquestapi.com *.twitter.com *.twimg.com *.aptrinsic.com 'unsafe-inline' 'unsafe-eval' data:; style-src * 'unsafe-inline' blob:; img-src * data:; font-src * data:; frame-src *;"
+        },
+        "cli": {}
+    },
+    "interleave": {
+        "dist\\bundle-src\\custom-widget.js": "cumulocity-datapoints-charting-widget-CustomWidget",
+        "dist/bundle-src/custom-widget.js": "cumulocity-datapoints-charting-widget-CustomWidget"
     }
 }
