+++ src/cumulocity-datapoints-charting-widget/cumulocity-datapoints-charting-widget.component.css
@@ -1,10 +1,4 @@
-/** @format */
-
-.chart-container {
-    display: block;
-    position: relative;
-    height: 90%;
-    /* border: 1px solid black; */
-    margin: 1rem;
+.cumulocity-datapoints-charting-widget-container {
+    display: grid;
+    justify-content: center;
 }
-
