+++ src/cumulocity-datapoints-charting-widget/cumulocity-datapoints-charting-widget.component.html
@@ -1,4 +1,4 @@
-<div class="chart-container" *ngIf="verifyConfig() else error">
+<div class="cumulocity-datapoints-charting-widget-container" *ngIf="verifyConfig() else error">
     <!-- <div style="display: block; width: 500px; height: 400px;"> -->
     <!-- <canvas baseChart [datasets]="chartData" [labels]="chartLabels" [options]="chartOptions" -->
     <ng-container *ngIf="chartData && chartData.length > 0">
@@ -32,7 +32,7 @@
     </ng-container>
 </div>
 <ng-template #error>
-    <div class="chart-container">
+    <div class="cumulocity-datapoints-charting-widget-container">
         <i class="fa fa-exclamation-circle fa-5x primary"></i>
         <h4>{{widgetHelper.getChartConfig().message}}</h4>
     </div>
