+++ src/cumulocity-datapoints-charting-widget/cumulocity-datapoints-charting-widget.config.component.css
@@ -1,10 +1,3 @@
-/** @format */
-
-h4 {
-    width: 100%;
-    text-align: center;
-}
-
 .configSection {
     display: grid;
     border: 1px solid rgba(0, 0, 0, 0.3);
@@ -21,11 +14,6 @@ h4 {
     display: inline-flex;
     align-content: baseline;
     align-items: center;
-}
-
-.seriesButton {
-    width: 100%;
-    text-align: left !important;
 }
 
 .help-icon {
