+++ src/cumulocity-datapoints-charting-widget/cumulocity-datapoints-charting-widget.module.ts
@@ -1,43 +1,54 @@
-/** @format */
+/**
+ * /*
+ * Copyright (c) 2019 Software AG, Darmstadt, Germany and/or its licensors
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @format
+ */
 
 import { CoreModule, HOOK_COMPONENTS } from "@c8y/ngx-components";
-import { CumulocityDataPointsChartingWidgetConfig as CumulocityDataPointsChartingWidgetConfig } from "./cumulocity-datapoints-charting-widget-config.component";
-import { CumulocityDataPointsChartingWidget as CumulocityDataPointsChartingWidget } from "./cumulocity-datapoints-charting-widget.component";
+import { CumulocityDatapointsChartingWidgetConfig } from "./cumulocity-datapoints-charting-widget.config.component";
+import { CumulocityDatapointsChartingWidget } from "./cumulocity-datapoints-charting-widget.component";
 import { NgModule } from "@angular/core";
+import { HttpClientModule } from "@angular/common/http";
 import { NgSelectModule } from "@ng-select/ng-select";
-import { ColorPickerModule } from "ngx-color-picker";
 import { ChartsModule } from "ng2-charts";
-//import { HttpClientModule } from "@angular/common/http";
-
-// This will import css from the styles folder (Note: will be applied globally, not scoped to the module/components)
-import "~styles/index.css";
-
-// You can also import css from a module
-// import 'some-module/styles.css'
 
 @NgModule({
-    imports: [CoreModule, NgSelectModule, ColorPickerModule, ChartsModule],
-    declarations: [CumulocityDataPointsChartingWidget, CumulocityDataPointsChartingWidgetConfig],
-    entryComponents: [CumulocityDataPointsChartingWidget, CumulocityDataPointsChartingWidgetConfig],
+    imports: [CoreModule, HttpClientModule, NgSelectModule, ChartsModule],
+    declarations: [CumulocityDatapointsChartingWidget, CumulocityDatapointsChartingWidgetConfig],
+    entryComponents: [CumulocityDatapointsChartingWidget, CumulocityDatapointsChartingWidgetConfig],
     providers: [
-        // Connect the widget to Cumulocity via the HOOK_COMPONENT injection token
         {
             provide: HOOK_COMPONENTS,
             multi: true,
             useValue: {
-                id: "global.presales.CumulocityDataPointsCharting.widget",
+                id: "com.softwareag.globalpresales.cumulocity.datapoints.charting.widget",
                 label: "Data Points Charting",
                 description: "Graph measurements and statistics about measurements",
-                component: CumulocityDataPointsChartingWidget,
-                configComponent: CumulocityDataPointsChartingWidgetConfig,
-                previewImage: require("~styles/previewImage.png"),
+                component: CumulocityDatapointsChartingWidget,
+                configComponent: CumulocityDatapointsChartingWidgetConfig,
+                previewImage: require("@widget-assets/img-preview.png"),
                 data: {
-                    settings: {
-                        noDeviceTarget: true,
+                    ng1: {
+                        options: { noDeviceTarget: false, noNewWidgets: false, deviceTargetNotRequired: false, groupsSelectable: true },
                     },
                 },
             },
         },
     ],
 })
-export class CumulocityDataPointsChartingWidgetModule {}
+export class CumulocityDatapointsChartingWidgetModule { }
