+++ styles/index.less
@@ -1,0 +1,9 @@
+/** @format */
+
+@import "~@c8y/style/extend.less";
+@import "~@ng-select/ng-select/themes/default.theme.css";
+
+canvas {
+    width: 100% !important;
+    height: 100% !important;
+}
