+++ tsconfig.json
@@ -1,22 +1,28 @@
 {
-  "compileOnSave": false,
-  "compilerOptions": {
-    "baseUrl": "./src",
-    "outDir": "./dist/out-tsc",
-    "sourceMap": true,
-    "declaration": false,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "target": "es5",
-    "downlevelIteration": true,
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-    "lib": [
-      "dom",
-      "es2016"
-    ]
-  }
+    "compileOnSave": false,
+    "compilerOptions": {
+        "paths": {
+            "@widget-assets/*": [
+                "./src/cumulocity-datapoints-charting-widget/assets/*",
+                "./dist/widget-library/src/cumulocity-datapoints-charting-widget/assets"
+            ]
+        },
+        "baseUrl": "./",
+        "outDir": "./dist/out-tsc",
+        "sourceMap": true,
+        "declaration": false,
+        "module": "esnext",
+        "moduleResolution": "node",
+        "emitDecoratorMetadata": true,
+        "experimentalDecorators": true,
+        "target": "es5",
+        "downlevelIteration": true,
+        "typeRoots": [
+            "node_modules/@types"
+        ],
+        "lib": [
+            "dom",
+            "es2016"
+        ]
+    }
 }
